Add SMD courtyard based keepout volumes
TH shells can interfere with SMD components due to their thickness.
To prevent this, we create keepout volumes for the SMD components.
Keepout volumes take the courtyard, and make it as tall as the
SMD component, based on the 3D model. The actual volume is more than
this - specifically a "gap" and a "clearance" value can be applied.

These values can be specified globally, or on a per component basis.
OpenSCAD preview shows the keepout volumes in pink using the debug
modifier "#"