courtyard shell, param tuning in oscad customizer
Courtyard shell type is implemented. It is now possible to tweak all
component parameters in OpenSCAD customizer at a per-component level:
- insertion direction
- shell thickness
- shell gap
- shell clearance from PCB
- include/exclude in model

There could be lots of refs on a board. Largest refs (by courtyard area)
are shown first in the customizer so that it's easier to find them.

This is quite nice, but it also means that the global values no longer
work. How do we deal with this ? Customizer does not allow cross
dependency between parameters.  So, to propagate values consistentyly,
we can at best aim to make it additive. This needs to be fixed in
the next few commits.

After thinking a bit about this, I think the way this should work is:
- global values will apply for all footprint types
- value overrides can be done on a per-footprint basis
- deltas can be applied on a per-component basis

As an example, we can have a 0.8 mm shell thickness.  For a particular
footprint type (say a Pico) this can be 1.2 mm. Let's say there are
two Picos on the board.  For one of them, we can choose a -0.4 value,
giving it a shell thickness of 0.8mm, for the other, we can set a
0.4 value, giving it a thickness of 1.6mm.