Beautified preview in oscad,with show/hide options
PCB, component holders and SMD keepout volumes can be independently
visualized.  These are clearly shown in different colors. It's not
perfect, but it's pretty good!

oscad background operator % is used as it allows color control, without
adding to output geometry. The debug modifier # has a fixed color, so
not used.

PCB is centered around the origin now to aid 3D interaction.
This is not applied if --keep-orientation is used.

Base can be switched between mesh/solid.