New "griddish" base type
This is a grid that includes all the footprint center. Basically,
it draws an X and a Y line on the base, for every footprint included
in the output.

This gives it a grid like appearance. The lines also likely pass
through the shell X & Y axis.

Delaunay can cause strange print moves, but look cool, depending on
who you ask.

Griddish has an advantage in that it dynamically updates in openscad,
unlike delaunay that is computed when gen-jig is run. This may make
it more useful for component fitting use cases too.