Hack-in-progress: tight shells
convex hulls works quite well for things like berg headers to generate
 "fitting" shells. But I want to generate concave shells for concave
parts - so that the cavity follows the model better. how to make that
happen ?

The ideal solution can be visualized as - offset the solid 3D model,
sweep it out along negative Z. This gives the subtraction volume, or
the holder for the component.

In this hack, I've used a modified z walk, based on the shell-zbins.
It works pretty well for something like a berg header - showing its
external outline perfectly. So far so good, but it crawls for complex
meshes with curved surfaces. Commit to save this work.

If you run this, you should see the "shadow" of the component when
the sun is at 12 o'clock