Better use of component display name in scad
Component's display name can provide extra info that is useful
for end users. E.g. something like "RPi HAT" is better than U1.
Display names are now shown, in addition to ref, if not same as
ref. Meaning - if the config file provides a distinct display
name that is not same as ref, it is accorded due importance!