Hack: fix issues in inner corners of fitting bins
Inner corners of fitting bins are created by subtracting volumes.
These serve as nice pockets, but also create new 3D printing challenges.
3D printers round off corners - this causes fitting issues.

A decent fix is avoiding tight corners.  This hack demonstrates this
in action. Basically, for every vertex in the polygon of the cutting
volumes, we add a circle the size of an FDM nozzle. This eliminates
the corner. Crucially, it keeps the corner clear of filament -
thus resulting in a good fit.

This also cuts into the outer shell, increasing the number of print
moves and also weakening it. Thus, 1.2 mm (3 nozzles) is the minimum
recommended shell thickness. That way, in places where the circle
cuts into ito - you'll still have 2 lines providing sufficient
shell width.