x & y line options for base. Faster prints ahoy
x_lines and y_lines are new options for the base pattern.
griddish was a combination of X & Y lines.  These are the
components.  They save effectively even more time and material.

Printing with x/y lines saves about 20 minutes of printing time,
down from 2 hrs 20 minutes to 2 hrs, compared to griddish.  And
simple lines make it easier for the print head also to move.

Since the lines connect the components all the way to the frame,
nothing else is actually required.  Visibility is also enhanced.