Corner avoidance with cuts for fitting shells
Corners were being cut earlier with circles.  This created issues as it
created variances in the shell width.

The change goes one better - we radially cut at corners - all the way to the
end of the shell. This saves material as well.  In the Z axis, the cuts
stretch all along the Z span of each bin.  This creates a very tiny
bridge for the next bin, but its so small that it won't be an issue.

Note that the cut volumes created by this corner avoidance cut through
all the shells. This is necessary as shells can potentially intrude into
other shells due to the thickness of their walls.