Initial implementation : Mounting hole spacer jig
Saves 50% material and 50% time compared to without, for the
AtiVEGA board.

Basically for a good height, the jig will just be printing
vertical shells. This can be substituted by using 3d printed
spacers. Base layer is still required.  The spacer is the
same shape as the mounting hole outlines.

Limitations: one or more extra base layers (near to topmost_z)
may need to be printed, if component ends around that place.
Need to add a policy for this.

Here, we use the first_layer_height as well. Shouldn't be a problem
as in most cases it is fair to assume this will be 0.2 mm

We also create shadow variables in scad.  Basically this jig
type overrides internal variables to do its job (mesh line thickness,
base height). This is done conditionally based on the state of the
checkbox