Output spacers in place, separable in slicer
This helps visualize exactly how it is supposed to be used.
User needs to load object in slicer and select "split to
objects" (Prusa Slicer) or similar option in other slicers.

This is also great to overlay onto the step file of the board.
Zooming in enough will show the gap.