Handling of base and wrapper in standalone mode
Footprints that have more than one 3d model (e.g. Pico, Mini360 etc)
need extra handling for the base and wrappers. In "jig" mode,
base is assumed, and things are supposed to be attached in some way
to the base.

In standalone mode, that doesn't happen.  So the base/wrappers have
to be generated additionally.