Fix: working courtyard shells in non-pcb mode
When "pure" footprints are loaded (i.e. without a PCB), somehow
the courtyards aren't returned. Shells were empty due to this.

I couldn't figure out from the KiCAD source code the exact reason.
But I was able to figure out that the drawings are available. So
I made my edge cuts code reusable and used that code to compute
the courtyards. With this, courtyard shells now work!