Mesh lines cut by volumes make islands. Avoid that
Cut volumes were earlier cutting into mesh lines. These are structures
designed to ensure the output is one piece. This is bad of course, as
it requires analysis of the result to figure out if something won't
work. Of course, base_thickness>0 can help by eliminating islands, but
that is potentially wasteful in terms of both material and time.

This change ensures mesh lines won't be cut into. So the structure will
remain one piece