feat:'minmesh' based on travelling salesman path
Travelling salesman path through all the component shells
and mounting holes is arguably a good alternative to delaunay
mesh - especially if supplemented by connecting each point to
the edge of the PCB.

That's exactly what this change does. TSP based path simplifies
printing as well, due to the lesser number of shapes on the
base.