Caching for "body shapes"
Each component on the board has a "footprint". It can also have
one/more 3D models, which decide the "shape" of the "body".

"body" is a new term created here. "body" has a "footprint" and
a "shape". KiCAD allows you to change the "shape" for every
"footprint", and it is possible that the user may knowingly or
unknowingly have slightly different shapes. The tool warns the
user in this case.

The "shape" of the "body" is decided by the 3d meshes, and their
transformation (translate, rotate and scale).

Caching takes into account any other factor that may be used to
derive hulls and bins - one example is the smallest area. Another
could be tesellation, but this isn't implemented yet.

Caching is introduced so that we can next perform more intense
calculations and cache the results.