Improved grid arrangement to look "natural"
When components are laid out in a grid, the layout is better if
max spans are computed along the minor dimension. If layout is
XY, then for every column the span is the max of the Y spans of
that column. This results in a more pleasing layout. Useful in
some ways - e.g. changing the wrapper thickness will yield more
predictable results.