feat:refs_do_not_process for SMD
This property is useful for SMD as well to selectively include/
exclude refs in the design (like TH).

Filter out refs that are in this list. Useful, among other reasons,
to deal with file format conversion/import issues (I am facing this
with the adafruit perma proto bonnet, see config.toml in commit)

Tested with:

./jigit \
  --enable-smd-shells \
  -i demo/adafruit-perma-proto-bonnet/adafruit-perma-proto-bonnet.kicad_pcb \
  -o demo/appb2.scad \
  -c demo/adafruit-perma-proto-bonnet/config.toml