Fix:Ensure mounting hole shells press PCB
If these stretch all the way till the PCB, there will be no
press action on the components.  Mounting hole shells must end a bit
short (same as SMD clearance) so that a small amount of force keeps
the components pressed. Note that this shouldn't be large - that
will distort the holding jig. 0.2 - 0.5 mm should be fine.