Handle missing models by excluding them from config
This looks like a much better solution. If a component has
no 3d models, it gets excluded rather than being retained
and causing errors later on.

This also ties in nicely with config file checking later on.
If a user specifies a component name that was excluded, it
is treated as an error. This way, any implicit assumptions
that the user has regarding the model are validated.