Pockets cut through PCB to avoid overhangs
Sometimes components can straddle the edge of the PCB, or be very
close to the edge.  When pockets don't cut through the PCB,
some plastic may hang in thin air near the edge. This change
ensures that all pockets (courtyard, wiggle, fitting and tight)
cut through the PCB. This is done by increasing the height of the cut
volume by pcb thickness, and making it begin at z=0.