force_mount option to flexibly place shells in Z
TH shell with min_z > 0 was taken to be not mounted. This was done
to process mounted modules such as the Pico on Thejas32.

force_mount ignores the min_z comparison and includes all models
in the shell. Doing so changes the meaning of clearance_from_pcb.
It now cuts from the top of the 3d model, if that has min_z > 0.
This allows us to cut the top of the shell independent of the
Z location, with a single parameter. If we didn't change this
interpretation, multiple components of the same type, with different
Z location would end up needing distinct values of clearance_from_pcb.

With this in place, we can have many components laid out arbitrarily
in Z. Paves the way to build interesting generic holders.