feat:implement trim volume
Trim volumes are planes that trim the jig above and below specified
Z values.  "below" refers to the usage orientation of the jig - i.e.
towards the base.

If the "below" volume is enabled, then the mesh is moved to that
location.

Trim volume locations are not checked. So, GIGO applies if you use
out of range. "below" must be smaller than max Z