solid base now works with trim below z
solid is a base style, but it wasn't moving with the trim_below_z
plane. This commit fixes this. Technically, this implements mesh
lines that are solid.

solid base uses the mounting hole frame as the shape for
component_fitting jigs. For TH_soldering, the PCB edge is used.

Couple of things to correct later (TODO):

  - solid_fill is not a base style technically; it's a jig style
  - MH frame or PCB should must be a policy
  - shell_style='support' must not overload shell_gap. Doing so
    complicates configuration file. Users need to maintain
    consistency across two or more config files.